# swagger.yaml
openapi: 3.0.3
info:
  title: Easy Planings Back API
  version: 1.0.0
  description: |
    A comprehensive API for managing group travel plans with real-time collaboration.

    ## Features
    - User authentication with JWT tokens
    - Create and manage trips with multiple members
    - Organize trips by days and activities
    - Invite other users to join your trips
    - Role-based access control (Owner/Member)

    ## Getting Started
    1. Register a new account using `/auth/register`
    2. Login to receive a JWT token using `/auth/login`
    3. Include the token in the `Authorization: Bearer <token>` header for protected endpoints

    ## Error Handling
    All error responses follow a consistent format with descriptive messages and error codes.
  contact:
    name: Easy Planings Support
    email: support@easyplanings.com
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.easyplanings.com
    description: Production server (when deployed)
...
security:
  - bearerAuth: []
paths:
...
  /{tripId}/days/{dayId}/activities/update-activities:
    post:
      tags: [Activities]
      summary: Create and/or update activities for a trip day
      operationId: updateActivities
      description: |
        Create new activities or update existing ones for a specific trip day.

        - To create a new activity, omit the `activityId` field
        - To update an existing activity, include its `activityId`
        - Both creation and update can be mixed in a single request
      parameters:
        - $ref: "#/components/parameters/TripIdPathParam"
        - $ref: "#/components/parameters/DayIdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivitiesRequest"
      responses:
        "200":
          description: Activities update summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateActivitiesResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Unauthorized or not a trip member
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UnauthorizedErrorBody"
                  - $ref: "#/components/schemas/CustomErrorBody"
        "404":
          description: Trip or day not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomErrorBody"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /{tripId}/days/{dayId}/activities/{activityId}/delete-activity:
    delete:
      tags: [Activities]
      summary: Delete an activity from a trip day
      operationId: deleteActivity
      description: Deletes a specific activity from a trip day. Only trip members can delete activities.
      parameters:
        - $ref: "#/components/parameters/TripIdPathParam"
        - $ref: "#/components/parameters/DayIdPathParam"
        - $ref: "#/components/parameters/ActivityIdPathParam"
      responses:
        "200":
          description: Activity deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Unauthorized or not a trip member
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UnauthorizedErrorBody"
                  - $ref: "#/components/schemas/CustomErrorBody"
        "404":
          description: Trip or day not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomErrorBody"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. 

        Get a token by logging in with `/auth/login` endpoint.
        Include the token in the Authorization header as: `Authorization: Bearer <your_token_here>`

        Tokens are valid for 7 days.
  parameters:
...
    TripIdPathParam:
      name: tripId
      in: path
      required: true
      description: Trip unique identifier (UUID)
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
...
    DayIdPathParam:
      name: dayId
      in: path
      required: true
      description: Trip day unique identifier (UUID)
      schema:
        type: string
        format: uuid
        example: "880e8400-e29b-41d4-a716-446655440333"
    ActivityIdPathParam:
      name: activityId
      in: path
      required: true
      description: Activity unique identifier (UUID)
      schema:
        type: string
        format: uuid
        example: "770e8400-e29b-41d4-a716-446655440222"
...
  responses:
    ValidationError:
      description: Request validation error - Fields did not meet the specified requirements
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorBody"
          examples:
            example:
              summary: Validation error example
              value:
                message:
                  - code: "invalid_type"
                    path: ["email"]
                    message: "Expected string, received undefined"
                  - code: "too_small"
                    path: ["password"]
                    message: "String must contain at least 8 character(s)"
...
    InternalServerError:
      description: Unexpected server error - An unexpected error occurred on the server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalServerErrorBody"
          examples:
            example:
              summary: Internal server error example
              value:
                message: "Internal server error"
                error: null
  schemas:
...
    UpdateActivitiesRequest:
      type: object
      required: [activities]
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityUpsert"
          description: Array of activities to create or update
    ActivityUpsert:
      type: object
      required: [title, description, location, startTime, endTime]
      properties:
        activityId:
          type: string
          format: uuid
          description: Existing activity id. If omitted, a new activity is created.
          example: "770e8400-e29b-41d4-a716-446655440222"
        title:
          type: string
          minLength: 1
          description: Activity name
          example: "Eiffel Tower Visit"
        description:
          type: string
          description: Detailed description of the activity
          example: "Visit the iconic Eiffel Tower with a guided tour"
        location:
          type: string
          minLength: 1
          description: Location of the activity
          example: "Eiffel Tower, Paris"
        startTime:
          type: string
          format: date-time
          description: Activity start time in ISO 8601 format
          example: "2026-06-01T18:00:00Z"
        endTime:
          type: string
          format: date-time
          description: Activity end time in ISO 8601 format
          example: "2026-06-01T21:00:00Z"
...
    Activity:
      type: object
      required:
        [
          id,
          title,
          startTime,
          endTime,
          createdAt,
          tripId,
          tripDayId,
          createdById,
        ]
      properties:
        id:
          type: string
          format: uuid
          description: Unique activity identifier
          example: "770e8400-e29b-41d4-a716-446655440222"
        title:
          type: string
          description: Activity name
          example: "Eiffel Tower Visit"
        description:
          type: string
          nullable: true
          description: Detailed description
          example: "Visit the iconic Eiffel Tower with a guided tour"
        location:
          type: string
          nullable: true
          description: Activity location
          example: "Eiffel Tower, Paris"
        startTime:
          type: string
          format: date-time
          description: Activity start time
          example: "2026-06-01T18:00:00Z"
        endTime:
          type: string
          format: date-time
          description: Activity end time
          example: "2026-06-01T21:00:00Z"
        createdAt:
          type: string
          format: date-time
          description: Activity creation timestamp
          example: "2026-06-01T10:00:00Z"
        tripId:
          type: string
          format: uuid
          description: Associated trip ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        tripDayId:
          type: string
          format: uuid
          description: Associated trip day ID
          example: "880e8400-e29b-41d4-a716-446655440333"
        createdById:
          type: string
          format: uuid
          description: ID of the user who created the activity
          example: "123e4567-e89b-12d3-a456-426614174000"
...
    UpdateActivitiesResult:
      type: object
      required: [created, updated]
      properties:
        created:
          type: integer
          minimum: 0
          description: Number of activities created
          example: 2
        updated:
          type: integer
          minimum: 0
          description: Number of activities updated
          example: 1
    UpdateActivitiesResponse:
      type: object
      required: [message, data]
      properties:
        message:
          type: string
          description: Response message
          example: "Successfully updated activities"
        data:
          $ref: "#/components/schemas/UpdateActivitiesResult"
    ActivityResponse:
      type: object
      required: [message, data]
      properties:
        message:
          type: string
          description: Response message
          example: "Activity successfully deleted"
        data:
          $ref: "#/components/schemas/Activity"
    ZodIssue:
      type: object
      description: Zod validation issue
      properties:
        code:
          type: string
          description: Error code
          example: "invalid_type"
        path:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          description: Path to the invalid field
          example: ["email"]
        message:
          type: string
          description: Error message
          example: "Invalid email"
    ValidationErrorBody:
      type: object
      required: [message]
      properties:
        message:
          type: array
          items:
            $ref: "#/components/schemas/ZodIssue"
          description: Array of validation errors
    UnauthorizedErrorBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Error message
          example: "Unauthorized access"
    CustomErrorPayload:
      type: object
      required: [name, statusCode]
      properties:
        name:
          type: string
          description: Error type
          example: "Error"
        statusCode:
          type: integer
          description: HTTP status code
          example: 404
        message:
          type: string
          description: Detailed error message
          example: "Trip not found"
    CustomErrorBody:
      type: object
      required: [message, error]
      properties:
        message:
          type: string
          description: Error summary
          example: "An error occurred"
        error:
          $ref: "#/components/schemas/CustomErrorPayload"
    InternalServerErrorBody:
      type: object
      required: [message, error]
      properties:
        message:
          type: string
          description: Error message
          example: "Internal server error"
        error:
          type: object
          nullable: true
          description: Error details (null in production)